Develop (#3)
* Генерация графика платежей теперь есть Говнокод, который работает.

* Много правок. Рефакторинг генерации кредитного предложения и графика платежей. Фиксы документации. Добавлены обработчики исключений.

* Пофиксил NPE в создании нового клиента (формирование списка кредитных предложений).

* Реализовал передачу полей "банк" и "тип кредита" в списке кредитных предложений.

* Реализовал передачу полей "банк" и "тип кредита" в списке кредитных предложений.

+++ src/main/java/ru/trainee/creditmanager/dto/loanOffer/LoanOfferResponseShortDTO.java
@@ -4,6 +4,8 @@ import lombok.*;
 import ru.trainee.creditmanager.dto.bank.BankResponseShortDTO;
 import ru.trainee.creditmanager.dto.creditType.CreditTypeResponseShortDTO;
 
+import java.math.BigDecimal;
+
 @EqualsAndHashCode
 @Builder
 @AllArgsConstructor
@@ -13,9 +15,10 @@ import ru.trainee.creditmanager.dto.creditType.CreditTypeResponseShortDTO;
 public class LoanOfferResponseShortDTO {
 
     Long id;
+//    BankResponseShortDTO bank;
     BankResponseShortDTO bank;
     CreditTypeResponseShortDTO creditType;
-    Integer sumOfCredit;
+    BigDecimal sumOfCredit;
     boolean isActive;
     boolean isAccepted;
 }
